Добавляет исправления
