+++ js/main.js
@@ -1,2 +1,6 @@
-import {renderingMiniatures} from "./rendering-miniatures.js";
-renderingMiniatures();
+import {renderMiniatures} from './render-miniatures.js';
+import {createPhotoDescriptions} from './data.js';
+
+const data = createPhotoDescriptions();
+
+renderMiniatures((data));
